chore: separate pinoHttp from logger
